Revert "Revert "Changed type of strikeprice from double to decimal to be consistent with the recent changes in tbloptions in TMLDB.""
This reverts commit 1af07f22379cbd726688eeacb915466c19beb37e.

+++ ICE_Import/ICE_Import/StoredProcs/SPOTable.sql
@@ -9,7 +9,7 @@ CREATE TABLE temp
     optionmonth char,
     optionmonthint int,
     optionyear int,
-    strikeprice float,
+    strikeprice decimal,
     callorput char,
     idinstrument int,
     expirationdate date,
